+++ RayTracing/Sphere.java
@@ -2,10 +2,10 @@ package RayTracing;
 
 public class Sphere {
 	final Vector center;
-	final int radius;
+	final float radius;
 	final Material material;
 
-	Sphere(int x, int y, int z, int radius, Material material) {
+	Sphere(float x, float y, float z, float radius, Material material) {
 		center = new Vector(x, y, z);
 		this.radius = radius;
 		this.material = material;
